Fix mismatching datatypes from API to mySQL and add DB actions.
mySQL returns a JSON object when querying a type of JSON object. This was
causing every entry that has a JSON object field to be labeled unique. The
API object was updated to match was mySQL returns.

Minor string issues were fixed, and null values are now handled properly.

DB actions were added for creating and deleting tables.