Adds second API configuration and setup.
Another API is needed because the sports data API intentionally
scrambles a lot of data for their free membership. However,
some of the data from that API is still very useful.

Combining the two APIs can allow us to store accurate and complete
data for games, stats, teams, and players.

The games, teams, and players have been setup and tested.