Fixes unique entry algorithm
When each task is preformed where API data gets transferred to
the database, a unique entry algorithm is preformed on the API
data.

This deciphers which entries from the API need to be updated in
the database, and also saves new entries to the database.

This also prevents duplicate entries in the database